requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dj-rest-auth==6.0.0
Django==5.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pika==1.3.2
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.2
pybase64==1.4.0
PyJWT==2.9.0
python-dotenv==1.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
